DistanceViewer memory removed.
This commit is the first commit of an experimental branch where I've removed the
memory function of the DistanceViewer. This function was used merely to be able
to view replays of the DistanceMapper mapping process, and whereas this was fun,
it wasn't very useful and ended up using ridiculous amounts of memory. This commit
has removed the function. However, I'm thinking of making it optional for those
that don't care about the memory.

I also added a small window to plot the change in the overal error during the mapping
process. This will need some work to be useful.