Fixed the delete function for DistanceViewers created from Phylip distances
Set up the connection so that the Phylip distance mappers are deleted when
the delete button is pressed. Also fixed the updatePositions point such that
no new point vectors are added to the parentPoints. This was causing a
segmentation fault during the delete procedure.. and is an indication of a
design fault somewhere... (mixed ownership of pointer issue probably).