Implemented a dimension squeezing algorithm for the Dimension Mapper
Instead of simply removing a dimension at the end of each mapping session,
and thus getting the really rather nice explosion scene, this algorithm
now gradually removes the space available for mapping in the last dimension
used.

This is achieved, simply by scaling all the distances in that dimension, by
a scaling factor which is linearly reduced towards 0. This should give a better
mapping as there should be less incidences of points becoming stuck in an
unsuitable position where they are unable to get to where they want to be.

The implementation in this algorithm is still somewhat ugly for two reasons:

1. It is not independent of the dimensionality reduction as it uses the currentDimNo
variable to work out which dimension to put the squeeze on.

2. It is clear that the higher dimensionality mapping requires much fewer iterations,
and we need some better way of setting the mapping for the dimensionality.

but for a 30 minute hack, it isn't half bad.