Serial Multidimensional squeezing cleanup (distanceMapper)
Changed the stress plotter so that it actually does something reasonable.
It now uses the maximum of the second half for scaling. This is still a bit
silly, but gives some sort of reasonable plot a the end.

Also made the stressplotter indicate the dimensionality shrinking so that
one can see the number of dimensions required to map the points in.

All a bit silly really, but it's ok.