Implemented a choice of multidimensional reduction methods
This version allows the user to either use a sudden (Starburst), a serial
or a parallel reduction in the dimensionality of the mapping space. This
was done mostly by introducing a new enum for the type of reduction and
redefining the DistanceMapper::reduceDimensionality function.

I also tried to clean up the code a little bit.