Minor cleanup and one bug squash
Changed the labels on some of the buttons to increase accuracy. Something
I should have done from the beginning. Note that the button labels now
reflect their functions better, but that the underlying code is still
mis-named (the change is from experiments to samples).

Also removed a simple bug in distanceViewer that would crash the program
if the start button was pressed whilst the mapping was still going on.
(This now does a wait() on the mapping process first, might be better to
change that to some sort of interrupt, but this is simple).