Minor cleanup of distanceMapper/stressPlotter and distanceMapper/distanceViewer
stressPlotter: Moved the code around to make the drawing function independent
of what it is drawn on. This to allow drawing to postcript or printers.

distanceViewer: Removed the non-functional replay button and moved the Start
button into it´s place. I don´t much like the new location, but maybe I´ll
get used to it.