Meaningless code cleanup
I made an effort to go through and remove a lot of the reasons for the
copious warnings that gcc throws out when compiling this project. Most
of those are related to comparison of signed vs. unsigned integers; and
most of those were in looping constructs (for int i=0;..etc). The main
reason I made the effort is in order for the more serious warnings that
actually can cause problems to be more visible. There are a few of those
around, mostly originating from a somewhat silly idea of mine to use
object memory addresses as identifiers, which might have been ok, except
that these will be a different size depdending on whether a computer is
32 or 64 bit (or whatever the future might bring). These don't actually
matter at the moment, since I never actually made any use of them. Still
at least the warnings for those more serious issues are not lost in a
mass of other warnings. To be completely frank, I'm still somewhat confused
as to when it is good to use unsigned and signed values, and this is
something that still bites me from time to time.

Naturally this caused some problems since I didn't perhaps do it carefully
enough. Hopefully I've caught all the problems caused by these changes,
but time will tell.