Added a record function for distanceMapper/pointDrawer.cpp
Added the ability to save a series of JPEGs of the mapping process
to directory. The directory selection is not very nice as it you
now have to select a new directory. This is seems to be limited
to the abilties of the QFileDialog; but it is probably primarily
related to my laziness. Since I do not expect this function to be
used much I don't want to spend too much time on it. The idea is
just to be able to make some movies of the mapping process for
demonstration purposes. This is not something that I think anyone
but me is likely to want to do.