Added functions to modify the current index, plus a bug fix
Added functions in clientWindow.cpp that allow the user to remove
the current, all above, or all below index entries. This should be
useful for preening statistically selected data sets.

I also removed a bug in the readState function (still clientWindow)
that got introduced when I was trying to reduce the number of
warnings for comparison of signed vs. unsigned integers.