Added a simple mechanism to remember sample selections
Added a simple set of functions and widgets that allow the user
to remember sets of samples (and their marks) selected by the
ExperimentChooser class. The current implementation is very simple
and does not handle order. It also cannot be saved to disk or to
the database and as such isn't a very good selection. See the
classes SampleMemoryWidget and SampleSelectionWidget for details.

It seems that I also removed a load of files that shouldn't have
been tracked by git. See complete diff for details.