Added a function that accesses the new expand index by genes function.
A selection of probe sets can now be expanded such that all probe sets
associated with the initial probe sets are included in the final probe
set selection.  This function obeys the chip selection and can thus be
used to inspect data from probe sets from other arrays and samples not
utilised in the initial selection specification.