Introduced multiple Raw Data plotwindows
Modified the code such that multiple Raw Data plotwindows can be
displayed simultaneously and be associated with independent sample
set selections.

These functions work independently from the previous sample set
memory functions implemented in the previous commits and can be
combined in all sorts of ways. Quite a few changes were introduced
and these have not been tested as thoroughly as I would like; there
may be some bugs.

Some functions (eg. exporting to html archives) are currently out of
order as it would not be clear as to which sample selection these
should be associated with. Well, to be honest, that's not really that
difficult, but I don't really like those functions anyway and so will
patch them up sometime in the future.