Fixed a couple of functions broken by the multiple plotwindows introduction
Fixed functions writeHtmlReport and writeCurrentProbeSet. These now go
through the plotWindows and finds the active one, and then use that
plotwindow to draw to a file.

Unfortunately the commit will have lots of new changes for the file Makefile.txt.
This file probably shouldn't be tracked by git as everytime a new Makefile.txt is
made these changes will appeear in the log. I've included the Makefiles in the
git repository as I like to have examples of working Makefiles (as I often need
to edit these). This is probably not the optimal way to do things and I'll
probably remove them at some point.